Start v1.5 development
The v1.5 is gonna be an intermediate version between v2 and v1, still relying on Action Cable protocol and its extensions, but providing AnyCable specific features. The v2 is planned to use a custom protocol.

We reserve v1.3 and v1.4 series for minor releases on top of the v1.2 (so we don't need to wait for the completion of v1.5 features)